+++ firmware/controllers/can/can_tx.cpp
@@ -19,6 +19,9 @@
 #if defined(MODULE_SHOCK_PRELOAD)
 #include "shock_preload.h"
 #endif
+#if defined(MODULE_EMP_PUMP)
+#include "emp_pump.h"
+#endif
 
 CanWrite::CanWrite(CanBusIndex bus)
 	: PeriodicController(bus == CanBusIndex::Bus0 ? "CAN TX 0" : "CAN TX 1", PRIO_CAN_TX, CAN_CYCLE_FREQ)
@@ -73,6 +76,9 @@ void CanWrite::PeriodicTask(efitick_t) {
 		// request on the CAN TX thread instead of blocking the main loop.
 		engine->module<ShockPreload>()->pollStatus();
 	}
+#endif
+#if defined(MODULE_EMP_PUMP)
+	engine->module<EmpPump>()->pollTx(m_bus);
 #endif
 
 #if EFI_WIDEBAND_FIRMWARE_UPDATE